feat: changed spconv and voxelization backbones to the ones from traveller59. cleaned part of the code but still needs some more. have not tested the node with the new changes

+++ autoware_lidar_bevfusion/include/autoware/tensorrt_plugins/quick_cumsum_cuda_plugin.hpp
@@ -36,10 +36,10 @@ namespace plugin
 
 struct QuickCumsumCudaParameters
 {
-  std::int64_t batch_size;
-  std::int64_t dimension;
-  std::int64_t height;
-  std::int64_t width;
+  std::int32_t batch_size;
+  std::int32_t dimension;
+  std::int32_t height;
+  std::int32_t width;
 };
 
 class QuickCumsumCudaPlugin : public IPluginV3,
